Relocate `diag` package under `helpers/diag`
