+++ helpers/diag/doc.go
@@ -1,0 +1,2 @@
+// Package diag provides diagnostics helpers for validator implementations.
+package diag
