+++ string/length_at_most.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/hashicorp/terraform-plugin-framework-validators/diag"
+	"github.com/hashicorp/terraform-plugin-framework-validators/helpers/diag"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 )
