--- validatordiag/doc.go
@@ -1,3 +1,0 @@
-// Package validatordiag provides diagnostics helpers for validator
-// implementations.
-package validatordiag
